+++ ICD.Connect.Misc.CrestronPro/Ports/ComPort/ComPortAdapter.cs
@@ -6,6 +6,7 @@ using ICD.Common.Utils.Services.Logging;
 using ICD.Connect.API.Nodes;
 using ICD.Connect.Devices.Extensions;
 using ICD.Connect.Misc.CrestronPro.Devices;
+using ICD.Connect.Misc.CrestronPro.Extensions;
 using ICD.Connect.Misc.CrestronPro.Utils;
 using ICD.Connect.Protocol.Ports;
 using ICD.Connect.Protocol.Ports.ComPort;
@@ -179,7 +180,7 @@ namespace ICD.Connect.Misc.CrestronPro.Ports.ComPort
 		protected override bool GetIsOnlineStatus()
 		{
 #if SIMPLSHARP
-			return m_Port != null;
+			return m_Port != null && m_Port.GetIsRegisteredAndParentOnline();
 #else
             return false;
 #endif
@@ -211,6 +212,10 @@ namespace ICD.Connect.Misc.CrestronPro.Ports.ComPort
 				return;
 
 			port.SerialDataReceived += PortSerialDataReceived;
+
+			GenericBase parent = port.Parent as GenericBase;
+			if (parent != null)
+				parent.OnlineStatusChange += ParentOnOnlineStatusChange;
 		}
 
 		/// <summary>
@@ -223,6 +228,10 @@ namespace ICD.Connect.Misc.CrestronPro.Ports.ComPort
 				return;
 
 			port.SerialDataReceived -= PortSerialDataReceived;
+
+			GenericBase parent = port.Parent as GenericBase;
+			if (parent != null)
+				parent.OnlineStatusChange -= ParentOnOnlineStatusChange;
 		}
 
 		/// <summary>
@@ -234,6 +243,11 @@ namespace ICD.Connect.Misc.CrestronPro.Ports.ComPort
 		{
 			PrintRx(args.SerialData);
 			Receive(args.SerialData);
+		}
+
+		private void ParentOnOnlineStatusChange(GenericBase currentDevice, OnlineOfflineEventArgs args)
+		{
+			UpdateCachedOnlineStatus();
 		}
 #endif